Resource algebra rebase (#434)
* Remove custom PCM combinator; rename storage protocol monoid naming

* Rename fraction namings

* Rename old int-based frac into counting permission.

* fmt vstd_extra
