+++ ostd/src/sync/rwlock.rs
@@ -3,7 +3,7 @@ use vstd::atomic_ghost::*;
 use vstd::cell::{self, pcell::*, CellId};
 use vstd::prelude::*;
 use vstd::resource::Loc;
-use vstd_extra::resource::ghost_resource::{csum::*, excl::*, tokens::*};
+use vstd_extra::resource::ghost_resource::{count::*, csum::*,  excl::*, tokens::*};
 use vstd_extra::sum::*;
 use vstd_extra::{prelude::*, resource};
 
@@ -43,10 +43,10 @@ exec const V_MAX_READ_RETRACT_FRACS: u64
 }
 
 /// The token reserved in the lock when the write permission is given out.
-type NoPerm<T> = Empty<PointsTo<T>>;
+type NoPerm<T> = EmptyCount<PointsTo<T>>;
 
 /// Half of the permission for read access, one for `RwLockUpgradeableGuard` and the other for all `RwLockReadGuard`s.
-type HalfPerm<T> = Frac<PointsTo<T>>;
+type HalfPerm<T> = Count<PointsTo<T>>;
 
 /// The permission for read access can be further split into `MAX_READER` pieces.
 type ReadPerm<T> = (HalfPerm<T>, OneLeftKnowledge<HalfPerm<T>, NoPerm<T>, 3>);
@@ -70,8 +70,8 @@ tracked struct RwPerms<T> {
     /// Tracks the number of live `RwLockReadGuard`s. If it is `Left`, it stores the remaining read permissions.
     /// If it is `Right`, it stores an empty token indicating the permission has been given out.
     read_guard_token: Sum<
-        FracResource<ReadPerm<T>, MAX_READER_U64>,
-        Empty<ReadPerm<T>, MAX_READER_U64>,
+        CountResource<ReadPerm<T>, MAX_READER_U64>,
+        EmptyCount<ReadPerm<T>, MAX_READER_U64>,
     >,
 }
 
@@ -369,7 +369,7 @@ impl<T, G> RwLock<T, G> {
         proof {
             lemma_consts_properties();
         }
-        let tracked mut frac_perm = Frac::<PointsTo<T>>::new(perm);
+        let tracked mut frac_perm = Count::<PointsTo<T>>::new(perm);
         let tracked read_half_cell_perm = frac_perm.split(1int);
         let ghost frac_id = frac_perm.id();
         let tracked mut core_token = SumResource::alloc_left(frac_perm);
@@ -377,7 +377,7 @@ impl<T, G> RwLock<T, G> {
         let tracked upread_retract_token = UniqueToken::alloc(());
         let tracked upreader_guard_token = core_token.split_one_left_owner();
         let tracked left_token = core_token.split_one_left_knowledge();
-        let tracked read_guard_token = FracResource::<ReadPerm<T>, MAX_READER_U64>::alloc(
+        let tracked read_guard_token = CountResource::<ReadPerm<T>, MAX_READER_U64>::alloc(
             (read_half_cell_perm, left_token),
         );
         let ghost v_id = RwId {
@@ -469,7 +469,7 @@ impl<T  /*: ?Sized*/ , G: SpinGuardian> RwLock<T, G> {
     #[verus_spec]
     pub fn try_read(&self) -> Option<RwLockReadGuard<'_, T, G>> {
         proof_decl!{
-            let tracked mut read_token: Option<Frac<ReadPerm<T>,MAX_READER_U64>> = None;
+            let tracked mut read_token: Option<Count<ReadPerm<T>,MAX_READER_U64>> = None;
             let tracked mut retract_read_token: Option<Token<V_MAX_READ_RETRACT_FRACS>> = None;
         }
         proof!{
@@ -718,7 +718,7 @@ unsafe impl<T: Sync, G: SpinGuardian> Sync for RwLockUpgradeableGuard<'_, T, G> 
 pub struct RwLockReadGuard<'a, T /*: ?Sized*/, G: SpinGuardian> {
     guard: G::ReadGuard,
     inner: &'a RwLock<T, G>,
-    v_token: Tracked<Frac<ReadPerm<T>, MAX_READER_U64>>,
+    v_token: Tracked<Count<ReadPerm<T>, MAX_READER_U64>>,
 }
 
 /*
@@ -939,7 +939,7 @@ impl<T  /*: ?Sized*/ , G: SpinGuardian> RwLockWriteGuard<'_, T, G> {
                 g.upreader_guard_token = Some(upreader_guard_token);
                 let tracked left_token = g.core_token.split_one_left_knowledge();
                 let tracked read_guard_empty = g.read_guard_token.tracked_take_right();
-                let tracked read_guard_token = FracResource::alloc_from_empty(
+                let tracked read_guard_token = CountResource::alloc_from_empty(
                     read_guard_empty,
                     (read_half_cell_perm, left_token),
                 );
