+++ ostd/src/sync/rwmutex.rs
@@ -4,7 +4,7 @@ use vstd::cell::{self, pcell::*, CellId};
 use vstd::prelude::*;
 use vstd::resource::Loc;
 use vstd_extra::auxiliary::pcell_borrow_mut;
-use vstd_extra::resource::ghost_resource::{csum::*, excl::*, tokens::*};
+use vstd_extra::resource::ghost_resource::{count::*, csum::*, excl::*, tokens::*};
 use vstd_extra::sum::*;
 
 use core::{
@@ -35,9 +35,9 @@ exec const V_MAX_READ_RETRACT_FRACS: u64
     (MAX_READER_MASK + 1) as u64
 }
 
-type NoPerm<T> = Empty<PointsTo<T>>;
+type NoPerm<T> = EmptyCount<PointsTo<T>>;
 
-type HalfPerm<T> = Frac<PointsTo<T>>;
+type HalfPerm<T> = Count<PointsTo<T>>;
 
 type ReadPerm<T> = (HalfPerm<T>, OneLeftKnowledge<HalfPerm<T>, NoPerm<T>, 3>);
 
@@ -47,8 +47,8 @@ tracked struct RwPerms<T> {
     upread_retract_token: Option<UniqueToken>,
     upreader_guard_token: Option<OneLeftOwner<HalfPerm<T>, NoPerm<T>, 3>>,
     read_guard_token: Sum<
-        FracResource<ReadPerm<T>, MAX_READER_U64>,
-        Empty<ReadPerm<T>, MAX_READER_U64>,
+        CountResource<ReadPerm<T>, MAX_READER_U64>,
+        EmptyCount<ReadPerm<T>, MAX_READER_U64>,
     >,
 }
 
@@ -303,7 +303,7 @@ impl<T> RwMutex<T> {
         proof {
             lemma_consts_properties();
         }
-        let tracked mut frac_perm = Frac::<PointsTo<T>>::new(perm);
+        let tracked mut frac_perm = Count::<PointsTo<T>>::new(perm);
         let tracked read_half_cell_perm = frac_perm.split(1int);
         let ghost frac_id = frac_perm.id();
         let tracked mut core_token = SumResource::alloc_left(frac_perm);
@@ -311,7 +311,7 @@ impl<T> RwMutex<T> {
         let tracked upread_retract_token = UniqueToken::alloc(());
         let tracked upreader_guard_token = core_token.split_one_left_owner();
         let tracked left_token = core_token.split_one_left_knowledge();
-        let tracked read_guard_token = FracResource::<ReadPerm<T>, MAX_READER_U64>::alloc(
+        let tracked read_guard_token = CountResource::<ReadPerm<T>, MAX_READER_U64>::alloc(
             (read_half_cell_perm, left_token),
         );
         let ghost v_id = RwId {
@@ -384,7 +384,7 @@ impl<T /*: ?Sized*/> RwMutex<T> {
     #[verus_spec]
     pub fn try_read(&self) -> Option<RwMutexReadGuard<'_, T>> {
         proof_decl! {
-            let tracked mut read_token: Option<Frac<ReadPerm<T>, MAX_READER_U64>> = None;
+            let tracked mut read_token: Option<Count<ReadPerm<T>, MAX_READER_U64>> = None;
             let tracked mut retract_read_token: Option<Token<V_MAX_READ_RETRACT_FRACS>> = None;
         }
         proof! {
@@ -594,7 +594,7 @@ unsafe impl<T:   /*: ?Sized +*/ Sync> Sync for RwMutexUpgradeableGuard<'_, T> {
 #[must_use]
 pub struct RwMutexReadGuard<'a, T  /*: ?Sized*/ > {
     inner: &'a RwMutex<T>,
-    v_token: Tracked<Frac<ReadPerm<T>, MAX_READER_U64>>,
+    v_token: Tracked<Count<ReadPerm<T>, MAX_READER_U64>>,
 }
 
 impl<'a, T> RwMutexReadGuard<'a, T> {
@@ -753,7 +753,7 @@ impl<'a, T  /*: ?Sized*/ > RwMutexWriteGuard<'a, T> {
                     upgrade_guard_token = Some(g.core_token.split_one_left_owner());
                     let tracked left_token = g.core_token.split_one_left_knowledge();
                     let tracked read_guard_empty = g.read_guard_token.tracked_take_right();
-                    let tracked read_guard_token = FracResource::alloc_from_empty(
+                    let tracked read_guard_token = CountResource::alloc_from_empty(
                         read_guard_empty,
                         (read_half_cell_perm, left_token),
                     );
@@ -826,7 +826,7 @@ impl<'a, T  /*: ?Sized*/ > RwMutexWriteGuard<'a, T> {
                 g.upreader_guard_token = Some(upreader_guard_token);
                 let tracked left_token = g.core_token.split_one_left_knowledge();
                 let tracked read_guard_empty = g.read_guard_token.tracked_take_right();
-                let tracked read_guard_token = FracResource::alloc_from_empty(
+                let tracked read_guard_token = CountResource::alloc_from_empty(
                     read_guard_empty,
                     (read_half_cell_perm, left_token),
                 );
