+++ vstd_extra/src/resource/ghost_resource/excl.rs
@@ -1,8 +1,7 @@
 //！ Exclusive ghost resource.
 use vstd::modes::tracked_swap;
 use vstd::prelude::*;
-use vstd::resource::algebra::Resource;
-use vstd::resource::algebra::ResourceAlgebra;
+use vstd::resource::algebra::{Resource, ResourceAlgebra};
 use vstd::resource::exclusive::ExclusiveRA;
 use vstd::resource::Loc;
