+++ vstd_extra/src/resource/ghost_resource/mod.rs
@@ -1,4 +1,5 @@
 //! Ghost resources available for use in proofs.
+pub mod count;
 pub mod csum;
 pub mod excl;
 pub mod tokens;
