+++ vstd_extra/src/resource/mod.rs
@@ -1,4 +1,3 @@
 //! Resource algebras based on Iris PCM and Leaf storage-protocol.
 pub mod ghost_resource;
-pub mod pcm;
 pub mod storage_protocol;
