+++ vstd_extra/src/resource/storage_protocol/mod.rs
@@ -1,6 +1,6 @@
 //! Aster-specific Leaf-style storage protocols.
 //!
 //! The generic storage-protocol API itself now comes from `vstd::resource::storage_protocol`.
-pub use vstd::resource::storage_protocol::*;
-
 pub mod csum;
+pub mod excl;
+pub mod hybrid_product;
